Add rubocop disable for unused parameter
The underscore was used to indicate that the parameter is not used in
the method. This is a common pattern in Ruby and is used to indicate
that the parameter is not needed for the method to function correctly.

However, this does not work with keyword arguments, so this change
disables this rubocop test instead.